Uploading the 5th homework
This is the 5th homework from my class and is designed around collecting data from GitHub.
In this I determine the number of people who have contributed to the apache hadoop public repo, the number of repos they have contributed, and the number of commits they have made to different repos.
Also looks at the difference between the first and last (100th) commit to the hadoop repo